requirements.txt
Flask
flask-cors
earthengine-api
firebase-admin
python-dotenv
gunicorn
google-cloud-secret-manager
